requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.37
sqlmodel==0.0.22
pydantic==2.9.2
python-multipart==0.0.20
pyaml==25.5.0
pydantic-settings==2.8.0
slowapi==0.1.9
pytest-mock==3.14.0
pytest-asyncio==0.25.3
alembic==1.16.5
python-dotenv
pymysql==1.1.2
PyJWT==2.10.1
pwdlib[argon2]
cryptography==46.0.3
pydantic[email]
redis==7.0.1
rq==2.6.0
